linting and small fixes

+++ include/trace.hpp
@@ -2,8 +2,8 @@
 #define TRACE_HPP
 
 #include <cstdint>
-#include <iomanip>
 #include <filesystem>
+#include <iomanip>
 #include <list>
 #include <ostream>
 #include <utility>
@@ -22,15 +22,16 @@ class Trace {
           cost(cost),
           averageDuration(averageDuration) {}
 
-    size_t getTimeStep() const noexcept;
+    uint64_t getTimeStep() const noexcept;
     size_t getNumberOfRequests() const noexcept;
     size_t getNumberOfOngoingSimulations() const noexcept;
     size_t getAvailableParkingSpots() const noexcept;
+
     double getCost() const noexcept;
     double getAverageDuration() const noexcept;
 
    private:
-   friend struct glz::meta<Trace>;
+    friend struct glz::meta<Trace>;
 
     uint64_t timestep;
     size_t numberOfRequests;
@@ -48,13 +49,10 @@ using Traces = std::list<Trace>;
 template <>
 struct glz::meta<palloc::Trace> {
     using T = palloc::Trace;
-    static constexpr auto value =
-        glz::object("timestep", &T::timestep,
-                    "number_of_requests", &T::numberOfRequests,
-                    "number_of_ongoing_simulations", &T::numberOfOngoingSimulations,
-                    "available_parking_spots", &T::availableParkingSpots,
-                    "cost", &T::cost,
-                    "average_duration", &T::averageDuration);
+    static constexpr auto value = glz::object(
+        "timestep", &T::timestep, "number_of_requests", &T::numberOfRequests,
+        "number_of_ongoing_simulations", &T::numberOfOngoingSimulations, "available_parking_spots",
+        &T::availableParkingSpots, "cost", &T::cost, "average_duration", &T::averageDuration);
 };
 
 #endif
